editado

+++ ejercicio3.cpp
@@ -6,9 +6,9 @@
 using namespace std;
 
 int main(){
-	int a;
-	int b;
-	int c;
+	float a;
+	float b;
+	float c;
 	
 	cout<< "ingrese a:\n";
 	cin>>a;
@@ -17,8 +17,8 @@ int main(){
 	cout<<"ingrese c:\n";
 	cin>>c;
 	
-	int x1 = ((-1*b+(sqrt((pow(b,2)-4*a*b))))/(2*a));
-	int x2 = ((-1*b-(sqrt((pow(b,2)-4*a*b))))/(2*a));
+	float x1 = ((-1*b+(sqrt((pow(b,2)-4*a*b))))/(2*a));
+	float x2 = ((-1*b-(sqrt((pow(b,2)-4*a*b))))/(2*a));
 	
 	cout<< "x1 vale:" <<x1;
 	cout<< "x2 vale:" << x2;
